requirements.txt
# requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.3
pytest==7.4.0
black==23.7.0
pylint==2.17.5
logging==0.4.9.6
pyarrow==18.1.0